+++ tests/Swarrot/Processor/Doctrine/ConnectionProcessorTest.php
@@ -3,9 +3,8 @@
 namespace Swarrot\Tests\Processor\Doctrine;
 
 use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Connections\MasterSlaveConnection;
-use Doctrine\DBAL\DBALException as DBAL2Exception;
-use Doctrine\DBAL\Exception as DBAL3Exception;
+use Doctrine\DBAL\Connections\PrimaryReadReplicaConnection;
+use Doctrine\DBAL\Exception as DBALException;
 use Doctrine\DBAL\Platforms\SqlitePlatform;
 use Doctrine\Persistence\ConnectionRegistry;
 use PHPUnit\Framework\TestCase;
@@ -37,14 +36,12 @@ class ConnectionProcessorTest extends TestCase
             $connectionProphecy->close()->shouldNotBeCalled();
             $connections[] = $connectionProphecy->reveal();
 
-            $connectionProphecy = $this->prophesizeMasterSlaveConnection();
-            $connectionProphecy->isConnectedToMaster()->willReturn(false);
+            $connectionProphecy = $this->prophesizePrimaryReplicaConnection();
             $connectionProphecy->isConnectedToPrimary()->willReturn(false);
             $connectionProphecy->close()->shouldNotBeCalled();
             $connections[] = $connectionProphecy->reveal();
 
-            $connectionProphecy = $this->prophesizeMasterSlaveConnection();
-            $connectionProphecy->isConnectedToMaster()->willReturn(true);
+            $connectionProphecy = $this->prophesizePrimaryReplicaConnection();
             $connectionProphecy->isConnectedToPrimary()->willReturn(true);
             $connectionProphecy->close()->shouldBeCalled();
             $connections[] = $connectionProphecy->reveal();
@@ -80,14 +77,12 @@ class ConnectionProcessorTest extends TestCase
             $connectionProphecy->close()->shouldNotBeCalled();
             $connections[] = $connectionProphecy->reveal();
 
-            $connectionProphecy = $this->prophesizeMasterSlaveConnection();
-            $connectionProphecy->isConnectedToMaster()->willReturn(false);
+            $connectionProphecy = $this->prophesizePrimaryReplicaConnection();
             $connectionProphecy->isConnectedToPrimary()->willReturn(false);
             $connectionProphecy->close()->shouldNotBeCalled();
             $connections[] = $connectionProphecy->reveal();
 
-            $connectionProphecy = $this->prophesizeMasterSlaveConnection();
-            $connectionProphecy->isConnectedToMaster()->willReturn(true);
+            $connectionProphecy = $this->prophesizePrimaryReplicaConnection();
             $connectionProphecy->isConnectedToPrimary()->willReturn(true);
             $connectionProphecy->close()->shouldBeCalled();
             $connections[] = $connectionProphecy->reveal();
@@ -103,12 +98,8 @@ class ConnectionProcessorTest extends TestCase
         $processor->process($message, $options);
     }
 
-    public function testCloseTimedOutConnectionDbal2()
+    public function testCloseTimedOutConnectionDbal()
     {
-        if (class_exists('\Doctrine\DBAL\Exception')) {
-            $this->markTestSkipped('doctrine/dbal >= 3.0 is installed');
-        }
-
         $innerProcessorProphecy = $this->prophesize(ProcessorInterface::class);
         $innerProcessorProphecy->process(Argument::cetera())->willReturn(true);
 
@@ -120,35 +111,7 @@ class ConnectionProcessorTest extends TestCase
         $connectionProphecy = $this->prophesizeConnection();
         $connectionProphecy->isConnected()->willReturn(true);
         $connectionProphecy->getDatabasePlatform()->willReturn($databasePlatformProphecy->reveal());
-        $connectionProphecy->query($dummySql)->willThrow(new DBAL2Exception());
-        $connectionProphecy->close()->shouldBeCalled();
-
-        $options = [
-            'doctrine_ping' => true,
-            'doctrine_close_master' => true,
-        ];
-        $processor = new ConnectionProcessor($innerProcessorProphecy->reveal(), [$connectionProphecy->reveal()]);
-        $processor->process(new Message(), $options);
-    }
-
-    public function testCloseTimedOutConnectionDbal3()
-    {
-        if (!class_exists('\Doctrine\DBAL\Exception')) {
-            $this->markTestSkipped('doctrine/dbal >= 3.0 is not installed');
-        }
-
-        $innerProcessorProphecy = $this->prophesize(ProcessorInterface::class);
-        $innerProcessorProphecy->process(Argument::cetera())->willReturn(true);
-
-        $dummySql = 'SELECT 1';
-
-        $databasePlatformProphecy = $this->prophesize(SqlitePlatform::class);
-        $databasePlatformProphecy->getDummySelectSQL()->willReturn($dummySql);
-
-        $connectionProphecy = $this->prophesizeConnection();
-        $connectionProphecy->isConnected()->willReturn(true);
-        $connectionProphecy->getDatabasePlatform()->willReturn($databasePlatformProphecy->reveal());
-        $connectionProphecy->query($dummySql)->willThrow(new DBAL3Exception());
+        $connectionProphecy->query($dummySql)->willThrow(new DBALException());
         $connectionProphecy->close()->shouldBeCalled();
 
         $options = [
@@ -187,8 +150,8 @@ class ConnectionProcessorTest extends TestCase
         return $this->prophesize(Connection::class);
     }
 
-    private function prophesizeMasterSlaveConnection()
+    private function prophesizePrimaryReplicaConnection()
     {
-        return $this->prophesize(MasterSlaveConnection::class);
+        return $this->prophesize(PrimaryReadReplicaConnection::class);
     }
 }